requirements.txt
requests
python-gvm
py7zr
pytz